+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Changed
+- Changed API of HeaderMapper. Now it can handler mandatory and optional names.  
+  The constructor is now private and a Builder has been added.  #13
+
+
 ## [0.14.2] - 2022-01-15
 ### Added
 - Added `WorkbookWriter.addCellAndComment()`.